requirements.txt
distribute==0.6.14
envoy==0.0.2
github3.py==0.8.2
requests==2.2.1
uritemplate.py==0.3.0
wsgiref==0.1.2
